requirements.txt
keyboard==0.13.5
pyyaml==6.0.1
PyQt5>=5.15
pytest>=7.0
pytest-qt>=4.0
